+++ alternative_backend/workers/urls.py
@@ -2,9 +2,10 @@ from django.urls import path
 from rest_framework.routers import DefaultRouter
 from workers.views import (
 	WorkerViewSet,
-	WorkerWorkHistoryAPIView,
 	WorkerLoginAPIView,
 	WorkerTokenValidate,
+	WorkerLogoutAPIView,
+	WorkerWorkHistoryAPIView,
 )
 
 
@@ -15,5 +16,6 @@ app_router.register(r'workers', WorkerViewSet, basename='workers')
 urlpatterns = [
 	path(r'worker_scan/', WorkerWorkHistoryAPIView.as_view()),
 	path(r'login/', WorkerLoginAPIView.as_view()),
+	path(r'logout/', WorkerLogoutAPIView.as_view()),
 	path(r'validate_token/', WorkerTokenValidate.as_view()),
 ] + app_router.urls