+++ CHANGELOG.md
@@ -2,12 +2,18 @@ CHANGELOG
 =========
 
 
-
 All notable changes to this project will be documented in this file.
 
 
+[1.0.2] - 2020-01-06
+--------------------
+Added
 
-[0.0.1] - 2020-01-06
+	- disable redis persistance, flush on every start
+	- adding logout endpoint
+
+
+[1.0.1] - 2020-01-06
 --------------------
 
 Added